requirements.txt
langchain-google-genai==4.2.1
langgraph==1.0.10
PyMuPDF==1.27.1
pydantic==2.12.5
python-dotenv==1.2.2
tqdm==4.67.3